refactor: Reorganize Cloudinary file structure for improved organization
In this commit, we have rearranged the locations of the Cloudinary files by creating a module for it to provide better structure and organization within our codebase.

The changes include:

1. Files related to Cloudinary have been grouped together in a dedicated directory. This makes it easier for developers to find and manage all Cloudinary-related code.

2. Naming conventions for Cloudinary files have been standardized to improve clarity.
