+++ src/modules/Cloudinary/types/Response/Response.ts
