+++ src/modules/Cloudinary/types/Response/index.ts
